# document-engine/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/pages/{id}/signed-url": {
      "get": {
        "tags": [
          "page-controller"
        ],
        "operationId": "signedUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedUrl"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pages/{id}/render": {
      "get": {
        "tags": [
          "page-controller"
        ],
        "operationId": "render",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/v1/packages/{packageId}/engine-results": {
      "get": {
        "tags": [
          "engine-result-controller"
        ],
        "operationId": "engineResultHistory",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngineResultMetadataView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/packages/{packageId}/engine-results/{revision}": {
      "get": {
        "tags": [
          "engine-result-controller"
        ],
        "operationId": "revision",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Length": {
                "description": "Exact canonical byte length",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Quoted envelope SHA-256",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "private, no-store",
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Type-Options": {
                "description": "nosniff",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/vnd.pragmaticds.document-engine-result+json;version=1": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/packages/{packageId}/engine-result": {
      "get": {
        "tags": [
          "engine-result-controller"
        ],
        "operationId": "current",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Length": {
                "description": "Exact canonical byte length",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Quoted envelope SHA-256",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "private, no-store",
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Type-Options": {
                "description": "nosniff",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/vnd.pragmaticds.document-engine-result+json;version=1": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/packages/{id}/pages": {
      "get": {
        "tags": [
          "page-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PackagePagesView"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/files/{id}/signed-url": {
      "get": {
        "tags": [
          "package-controller"
        ],
        "operationId": "signedUrl_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedUrl"
                }
              }
            }
          }
        }
      }
    },
    "/v1/files/{id}/content": {
      "get": {
        "tags": [
          "package-controller"
        ],
        "operationId": "content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/download": {
      "get": {
        "tags": [
          "signed-download-controller"
        ],
        "operationId": "download",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/documents/{id}/pdf/signed-url": {
      "get": {
        "tags": [
          "document-pdf-controller"
        ],
        "operationId": "signedUrl_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedUrl"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{id}/history": {
      "get": {
        "tags": [
          "document-review-controller"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IssuedUrl": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EngineResultMetadataView": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "processingJobId": {
            "type": "string",
            "format": "uuid"
          },
          "parseGeneration": {
            "type": "integer",
            "format": "int32"
          },
          "materializedJobAttempt": {
            "type": "integer",
            "format": "int32"
          },
          "envelopeSchemaVersion": {
            "type": "string"
          },
          "sourceSetSha256": {
            "type": "string"
          },
          "provenanceSha256": {
            "type": "string"
          },
          "envelopeSha256": {
            "type": "string"
          },
          "envelopeSizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "reuseEligibility": {
            "type": "string",
            "enum": [
              "PARSE_ONCE_CURRENT_PACKAGE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PackagePagesView": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "format": "uuid"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageView"
            }
          }
        }
      },
      "PageView": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceFileId": {
            "type": "string",
            "format": "uuid"
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "packagePageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "widthPt": {
            "type": "number"
          },
          "heightPt": {
            "type": "number"
          },
          "rotation": {
            "type": "integer",
            "format": "int32"
          },
          "detectedRotation": {
            "type": "integer",
            "format": "int32"
          },
          "renderDpi": {
            "type": "integer",
            "format": "int32"
          },
          "hasRender": {
            "type": "boolean"
          },
          "textLayer": {
            "type": "string"
          },
          "blank": {
            "type": "boolean"
          },
          "duplicateOfPageId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceContentType": {
            "type": "string"
          }
        }
      },
...
      "HistoryEntry": {
        "type": "object",
        "properties": {
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldName": {
            "type": "string"
          },
          "groupKey": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "previousValue": {
            "type": "object"
          },
          "newValue": {
            "type": "object"
          },
          "reason": {
            "type": "string"
          },
          "decidedBy": {
            "type": "string",
            "format": "uuid"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HistoryResult": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}